implemente stripe subscriptions

+++ app/api/board/route.js
@@ -9,18 +9,25 @@ export async function POST(req) {
     const body = await req.json();
     if (!body) {
       return NextResponse.json(
-        { message: "Board name is required" },
+        { error: "Board name is required" },
         { status: 400 }
       );
     }
     const session = await auth();
     if (!session) {
-      return NextResponse.json({ message: "Not authorized" }, { status: 401 });
+      return NextResponse.json({ error: "Not authorized" }, { status: 401 });
     }
 
     await connectMongo();
 
     const user = await User.findById(session.user.id);
+
+    if (!user.hasAccess) {
+      return NextResponse.json(
+        { error: "Please subscribe first" },
+        { status: 403 }
+      );
+    }
 
     const board = await Board.create({
       userId: user._id,
@@ -43,7 +50,7 @@ export async function DELETE(req) {
 
     if (!boardId) {
       return NextResponse.json(
-        { message: "Board ID is required" },
+        { error: "Board ID is required" },
         { status: 400 }
       );
     }
@@ -51,19 +58,29 @@ export async function DELETE(req) {
     const session = await auth();
 
     if (!session) {
-      return NextResponse.json({ message: "Not authorized" }, { status: 401 });
+      return NextResponse.json({ error: "Not authorized" }, { status: 401 });
     }
 
     await connectMongo();
 
-    await Board.deleteOne({ _id: boardId, userId: session?.user?.id });
-
     const user = await User.findById(session?.user?.id);
 
+    if (!user.hasAccess) {
+      return NextResponse.json(
+        {
+          error: "Please subscribe first",
+        },
+        { status: 403 }
+      );
+    }
+
+    await Board.deleteOne({ _id: boardId, userId: session?.user?.id });
+
     user.boards = user.boards.filter((id) => id.toString() !== boardId);
+    user.save();
 
     return NextResponse.json(
-      { message: "Board deleted successfully" },
+      { error: "Board deleted successfully" },
       { status: 200 }
     );
   } catch (e) {
